rename(dialogs): dialogsディレクトリを作成・アバター用ダイアログをリネーム
