+++ src/app/settings/settings/settings.component.html
@@ -42,7 +42,7 @@
         <button
           mat-flat-button
           color="primary"
-          (click)="openImageChangeDialog()"
+          (click)="openUserAvaterDialog()"
         >
           アバターを変更する
         </button>
