+++ src/app/shared/dialogs/user-avatar/user-avater.component.scss
